==> 9_1_Data_Preprocessing_After_ReSampling_SWANSF.ipynb <==
# Creating NewFeatures and Normalization

# New Features 

# NewFeatures with Shuffle

def new_Features_pkl(start_partition, end_partition, data_dir, first_name, last_name, data, labels):

    number_of_new_features = 9
    num_attributes = 24
    num_timestamps = 60
    
    for i in range(start_partition-1,end_partition):
        
        each_partition = np.zeros((data[i].shape[0], num_timestamps*(num_attributes)))
        each_partition = np.array(data[i])
        new_partition = np.zeros((each_partition.shape[0], number_of_new_features*num_attributes))
        
        with tqdm(each_partition.shape[0]) as pbar:
            for j in range(0,each_partition.shape[0]):
                new_column = np.zeros((num_timestamps*num_attributes)) 
                new_column = each_partition[j,:]
                

                new_features = np.zeros(number_of_new_features*(num_attributes))

                for m in range(0,num_attributes):
                    
                    mean = np.mean(new_column[m*num_timestamps:(m+1)*num_timestamps])
                    new_features[((m)*number_of_new_features) + 0] = mean
                    median = np.median(new_column[m*num_timestamps:(m+1)*num_timestamps])
                    new_features[((m)*number_of_new_features) + 1] = median
                    std = np.std(new_column[m*num_timestamps:(m+1)*num_timestamps])
                    new_features[((m)*number_of_new_features) + 2] = std
                    
                    skewness = skew(new_column[m*num_timestamps:(m+1)*num_timestamps])
                    if skewness == np.nan:
                        skewness = new_partition[j-1, ((m)*number_of_new_features) + 3]
                    new_features[((m)*number_of_new_features) + 3] = skewness
                    
                    kurtosis_value = kurtosis(new_column[m*num_timestamps:(m+1)*num_timestamps])
                    if kurtosis_value == np.nan:
                        kurtosis_value = new_partition[j-1, ((m)*number_of_new_features) + 4]
                    new_features[((m)*number_of_new_features) + 4] = kurtosis_value
                    
                    indices = np.arange(num_timestamps)
                    weight_array = indices / num_timestamps
                    weighted_avg = np.average(new_column[m*num_timestamps:(m+1)*num_timestamps], weights=weight_array)
                    if weighted_avg == np.nan:
                        weighted_avg = new_partition[j-1, ((m)*number_of_new_features) + 5]
                    new_features[((m)*number_of_new_features) + 5] = weighted_avg
                    
                    last_value = new_column[(m+1)*num_timestamps-1]
                    new_features[((m)*number_of_new_features) + 6] = last_value
                    first_value = new_column[m*num_timestamps]
                    new_features[((m)*number_of_new_features) + 7] = first_value
                    
                    numerator = np.sum((new_column[m*num_timestamps:(m+1)*num_timestamps] - mean) * (indices - np.mean(indices)))
                    denominator = np.sum((new_column[m*num_timestamps:(m+1)*num_timestamps] - mean) ** 2)
                    slope = numerator / denominator
                    if slope == np.nan:
                        slope = new_partition[j-1, ((m)*number_of_new_features) + 8]
                    new_features[((m)*number_of_new_features) + 8] = slope
                
                    
                new_partition[j,:] = new_features
                
                pbar.update(1)
                

        print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(new_partition).any()))  
        X_train = new_partition
        Y_train = labels[i]


        num_samples = X_train.shape[0]
        shuffle_indices = np.random.permutation(num_samples)

        X_train_shuffled = X_train[shuffle_indices]
        Y_train_shuffled = Y_train[shuffle_indices]
    

        with open(data_dir + first_name + "Partition" + str(i+1) 
                       + last_name + ".pkl", 'wb') as f:
            pickle.dump(X_train_shuffled, f)

        with open(data_dir + first_name + "Partition" + str(i+1) 
                       + "_Labels" + last_name + ".pkl", 'wb') as f:
            pickle.dump(Y_train_shuffled, f)

# Loading data

## RUS_Tomek_Smote 

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_Smote_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_Smote_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_1_OUSampling_NewFeatures_KnnImputation/"
name_first_part = 'RUS_Tomek_Smote_'
name_last_part = '_OUSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## RUS_Tomek_Adasyn

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_Adasyn_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_Adasyn_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_1_OUSampling_NewFeatures_KnnImputation/"
name_first_part = 'RUS_Tomek_Adasyn_'
name_last_part = '_OUSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## Smote

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "Smote_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "Smote_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_2_OSampling_NewFeatures_KnnImputation/"
name_first_part = 'Smote_'
name_last_part = '_OSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## Adasyn 

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "Adasyn_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "Adasyn_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_2_OSampling_NewFeatures_KnnImputation/"
name_first_part = 'Adasyn_'
name_last_part = '_OSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## GNI

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "GaussianNoise_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "GaussianNoise_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_2_OSampling_NewFeatures_KnnImputation/"
name_first_part = 'GaussianNoise_'
name_last_part = '_OSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## GNI, Tomek, RUS 

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_GNI_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_GNI_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_1_OUSampling_NewFeatures_KnnImputation/"
name_first_part = 'RUS_Tomek_GNI_'
name_last_part = '_OUSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## TimeGAN

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "TimeGAN_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "TimeGAN_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_2_OSampling_NewFeatures_KnnImputation/"
name_first_part = 'TimeGAN_'
name_last_part = '_OSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)

## TimeGAN, Tomek, RUS 

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_TimeGAN_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_TimeGAN_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

from tqdm import tqdm
import pandas as pd
import numpy as np
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/7_1_OUSampling_NewFeatures_KnnImputation/"
name_first_part = 'RUS_Tomek_TimeGAN_'
name_last_part = '_OUSampling_WithoutC_NewFeatures_KnnImputation'


new_Features_pkl(1, 5, data_dir, name_first_part, name_last_part, X_train_array, Y_train_array)
