==> 10_4_BinaryClassification_After_ReSampling_Part1_SWANSF.ipynb <==
# Rocket, TSF, LSTM, RNN, GRU, 1D-CNN on Cocatenation (After Sampling)

# Train Test Split

## Stratified Cross validation (k = 5)

P1 Train and P2 Test, P1 Train and P3 Test, P1 Train and P4 Test, P1 Train and P5 Test
P2 Train and P3 Test, P2 Train and P4 Test, P2 Train and P5 Test
P3 Train and P4 Test, P3 Train and P5 Test
P4 Train and P5 Test

# Evaluation Metrics

# float("{:.2f}".format(13.949999999999999))

import warnings
warnings.filterwarnings('ignore')

def TSS(TP,TN,FP,FN):
    TSS_value = (TP / (TP + FN)) - (FP / (FP + TN))
    return TSS_value

def HSS1(TP,TN,FP,FN):
    HSS1_value = (2 * (TP * TN - FP * FN)) / ((TP + FN) * (FN + TN) + (TP + FP) * (FP + TN))
    return HSS1_value
    
def HSS2(TP,TN,FP,FN):
    HSS2_value = (2 * (TP * TN - FP * FN)) / ((TP + FP) * (FN + TN) + (TP + FN) * (FP + TN))
    return HSS2_value

def GSS(TP,TN,FP,FN):
    GSS_value = (TP - (TP + FP) * (TP + FN) / (TP + FP + FN + TN))
    return GSS_value

def Recall(TP,TN,FP,FN):
    Recall_value = (TP) / (TP + FN)
    return Recall_value

def FPR(TP,TN,FP,FN):
    fpr_value = (FP) / (FP + TN)
    return fpr_value

def Accuracy(TP,TN,FP,FN):
    accuracy_value = (TP + TN) / (TP + TN + FP + FN)
    return accuracy_value

def Precision(TP,TN,FP,FN):
    precision_value = (TP) / (TP + FP)
    return precision_value

# Loading the Final Datasets

## Concatenation 

#### RUS-Tomek_GNI

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
X_train_concat_ZM_RTS = []
Y_train_concat_ZM_RTS = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "RUS_Tomek_GNI_" +"Partition" + str(i+1) + "_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        X_train_concat_ZM_RTS.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_GNI_" + "Partition" + str(i+1) + "_Labels_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        Y_train_concat_ZM_RTS.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_concat_ZM_RTS[i]).any() or np.isinf(X_train_concat_ZM_RTS[i]).any()))

#### RUS-Tomek_TimeGAN

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
X_train_concat_ZM_RTA = []
Y_train_concat_ZM_RTA = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "RUS_Tomek_TimeGAN_" +"Partition" + str(i+1) + "_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        X_train_concat_ZM_RTA.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_TimeGAN_" + "Partition" + str(i+1) + "_Labels_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        Y_train_concat_ZM_RTA.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_concat_ZM_RTA[i]).any() or np.isinf(X_train_concat_ZM_RTA[i]).any()))

#### GNI

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
X_train_concat_ZM_S = []
Y_train_concat_ZM_S = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "GaussianNoise_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        X_train_concat_ZM_S.append(pickle.load(f))
    with open(data_dir + "GaussianNoise_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        Y_train_concat_ZM_S.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_concat_ZM_S[i]).any() or np.isinf(X_train_concat_ZM_S[i]).any()))

#### TimeGAN

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
X_train_concat_ZM_A = []
Y_train_concat_ZM_A = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "TimeGAN_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        X_train_concat_ZM_A.append(pickle.load(f))
    with open(data_dir + "TimeGAN_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation" +".pkl", 'rb') as f:
        Y_train_concat_ZM_A.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_concat_ZM_A[i]).any() or np.isinf(X_train_concat_ZM_A[i]).any()))

#### Raw Data

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_1_FinalData_Concatenation_KnnImputation/"

concat_data = []
concat_data_labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        concat_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(concat_data[i]).any() or np.isinf(concat_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        concat_data_labels.append(pickle.load(f))

#### Test Data

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_4_FinalData_Concatenation_LSBZM_KnnImputation/"
X_test_concat_ZM = []
Y_test_concat_ZM = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        X_test_concat_ZM.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_test_concat_ZM[i]).any() or np.isinf(X_test_concat_ZM[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_Concatenation_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        Y_test_concat_ZM.append(pickle.load(f))

# Useful Functions

def kfold_training(name, X_train, Y_train, X_test, Y_test, training_func, num):
    kfold = np.array([[1,2],[2,3],[3,4],[4,5]])

    metrics = []
    metrics_values = np.array([])
    
    for i in range(0, num):
        train_index = kfold[i,0]
        test_index = kfold[i,1]
        metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_test[test_index-1], Y_test[test_index-1])
        while (metrics_values[4] < 0.2):
            metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_test[test_index-1], Y_test[test_index-1])
        metrics.append(np.append(np.append(train_index, test_index), metrics_values))
    return metrics

# Training

## Rocket (RidgeClassifierCV) 

# ROCKET with RidgeClassifierCV

import numpy as np
from sklearn.linear_model import RidgeClassifierCV
from sktime.transformations.panel.rocket import Rocket
from sklearn.metrics import confusion_matrix
import joblib
from sktime.datatypes._panel._convert import from_2d_array_to_nested

def rocket_model(X_train, Y_train, X_test, Y_test, rocket_kernels=1500):

    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/rocket/"
    
    X_train = from_2d_array_to_nested(X_train)
    X_test = from_2d_array_to_nested(X_test)

    rocket = Rocket(num_kernels=rocket_kernels)
    rocket.fit(X_train)
    X_train_transform = rocket.transform(X_train)
        
    classifier = RidgeClassifierCV()
    classifier.fit(X_train_transform, Y_train)
    
    X_test_transform = rocket.transform(X_test)
    
    y_pred = classifier.predict(X_test_transform)
    
    print(str(X_train.shape)+': Rocket Classifier is Done! \n')

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "rocket_model.pkl")

    #loaded_rocket_model = joblib.load(data_dir + "rocket_model_sgd.pkl")
    
    return output_values

## TimeSeriesForest

# TimeSeriesForest

from sktime.classification.interval_based import TimeSeriesForestClassifier
import numpy as np
from sklearn.metrics import confusion_matrix
import joblib

def tsf_model(X_train, Y_train, X_test, Y_test, tsf_estimator=60):

    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/TSF/"
    
    
    tsf_classifier = TimeSeriesForestClassifier(min_interval=30, n_estimators=60)
    tsf_classifier.fit(X_train, Y_train)
    y_pred = tsf_classifier.predict(X_test)
    
    print(str(X_train.shape)+': TSF Classifier is Done! \n')
    

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "tsf_model.pkl")

    #loaded_rocket_model = joblib.load(data_dir + "tsf_model.pkl")
    
    return output_values

## ShapeDTW

# TimeSeriesForest

from sktime.classification.dictionary_based import BOSSEnsemble
import numpy as np
from sklearn.metrics import confusion_matrix
import joblib

def knnts_model(X_train, Y_train, X_test, Y_test):

    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/BOSSEnsemble/"
    
    
    knnts_classifier = BOSSEnsemble(max_ensemble_size=3)
    knnts_classifier.fit(X_train, Y_train)
    y_pred = knnts_classifier.predict(X_test)
    
    print(str(X_train.shape)+': BOSSEnsemble Classifier is Done! \n')
    

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "tsf_model.pkl")

    #loaded_rocket_model = joblib.load(data_dir + "tsf_model.pkl")
    
    return output_values

## LSTM

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import keras
from keras.models import Sequential
from keras.layers import LSTM, Dense, Dropout
from sklearn.metrics import confusion_matrix

def lstm_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"
    
    verbose, epochs, batch_size = 0, 15, 32
    n_timesteps, n_features = 60, 24

    model = Sequential()
    model.add(LSTM(120, input_shape=(n_timesteps,n_features)))
    model.add(Dropout(0.3))
    model.add(Dense(120, activation='relu'))
    model.add(Dense(2, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=keras.metrics.SpecificityAtSensitivity(sensitivity=0.98))
    
    model.fit(X_train, Y_train, epochs=epochs, batch_size=batch_size, verbose=verbose)
    
    best_threshold = 0.0
    best_tss = 0.0
    y_pred = model.predict(X_test)
    # evaluate model
    for i in range(1, 1000):

        threshold = i / 1000 # Adjust the threshold as needed
        y_pred_binary = (y_pred > threshold).astype(int)
        confusion = confusion_matrix(Y_test, y_pred_binary)
        tn, fp, fn, tp = confusion.ravel()
        tss = TSS(tp,tn,fp,fn)
        if tss > best_tss:
            best_tss = tss
            best_threshold = i / 1000
        
    
    print(str(X_train.shape)+': LSTM Classifier is Done! \n')


    threshold = best_threshold # Adjust the threshold as needed
    y_pred_binary = (y_pred > threshold).astype(int)
    confusion = confusion_matrix(Y_test, y_pred_binary)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "mlp_model.pkl")

    #loaded_mlp_model = joblib.load(data_dir + "mlp_model.pkl")
    
    return output_values

## 1D-CNN

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import keras
from keras.models import Sequential
from keras.layers import Conv1D, MaxPooling1D, Flatten, Dense, Dropout
from sklearn.metrics import confusion_matrix

def cnn_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"
    
    verbose, epochs, batch_size = 0, 15, 32
    n_timesteps, n_features = 60, 24
    
    model = Sequential()
    model.add(Conv1D(filters=64, kernel_size=5, activation='relu', input_shape=(n_timesteps,n_features)))
    model.add(Dropout(0.2))
    model.add(MaxPooling1D(pool_size=2))
    model.add(Conv1D(filters=128, kernel_size=5, activation='relu'))
    model.add(Dropout(0.3))
    model.add(MaxPooling1D(pool_size=2))
    model.add(Flatten())
    model.add(Dense(357, activation='relu'))
    model.add(Dense(128, activation='relu'))
    model.add(Dense(2, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=keras.metrics.SpecificityAtSensitivity(sensitivity=0.98))
    
    model.fit(X_train, Y_train, epochs=epochs, batch_size=batch_size, verbose=verbose)
    
    best_threshold = 0.0
    best_tss = 0.0
    y_pred = model.predict(X_test)
    # evaluate model
    for i in range(1, 1000):

        threshold = i / 1000 # Adjust the threshold as needed
        y_pred_binary = (y_pred > threshold).astype(int)
        confusion = confusion_matrix(Y_test, y_pred_binary)
        tn, fp, fn, tp = confusion.ravel()
        tss = TSS(tp,tn,fp,fn)
        if tss > best_tss:
            best_tss = tss
            best_threshold = i / 1000
        
    
    print(str(X_train.shape)+': CNN Classifier is Done! \n')


    threshold = best_threshold # Adjust the threshold as needed
    y_pred_binary = (y_pred > threshold).astype(int)
    confusion = confusion_matrix(Y_test, y_pred_binary)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "mlp_model.pkl")

    #loaded_mlp_model = joblib.load(data_dir + "mlp_model.pkl")
    
    return output_values

## RNN

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import keras
from keras.models import Sequential
from keras.layers import SimpleRNN, Dense, Dropout
from sklearn.metrics import confusion_matrix

def rnn_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"
    
    verbose, epochs, batch_size = 0, 15, 32
    n_timesteps, n_features = 60, 24
    
    model = Sequential()
    model.add(SimpleRNN(units=120, activation='relu', input_shape=(n_timesteps,n_features)))
    model.add(Dropout(0.3))
    model.add(Dense(120, activation='relu'))
    model.add(Dense(2, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=keras.metrics.SpecificityAtSensitivity(sensitivity=0.98))
    
    model.fit(X_train, Y_train, epochs=epochs, batch_size=batch_size, verbose=verbose)


    best_threshold = 0.0
    best_tss = 0.0
    y_pred = model.predict(X_test)
    # evaluate model
    for i in range(1, 1000):

        threshold = i / 1000 # Adjust the threshold as needed
        y_pred_binary = (y_pred > threshold).astype(int)
        confusion = confusion_matrix(Y_test, y_pred_binary)
        tn, fp, fn, tp = confusion.ravel()
        tss = TSS(tp,tn,fp,fn)
        if tss > best_tss:
            best_tss = tss
            best_threshold = i / 1000
        
    
    print(str(X_train.shape)+': RNN Classifier is Done! \n')


    threshold = best_threshold # Adjust the threshold as needed
    y_pred_binary = (y_pred > threshold).astype(int)
    confusion = confusion_matrix(Y_test, y_pred_binary)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "mlp_model.pkl")

    #loaded_mlp_model = joblib.load(data_dir + "mlp_model.pkl")
    
    return output_values

## GRU 

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import keras
from keras.models import Sequential
from keras.layers import GRU, Dense, Dropout
from sklearn.metrics import confusion_matrix

def gru_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"
    
    verbose, epochs, batch_size = 0, 15, 32
    n_timesteps, n_features = 60, 24
    
    model = Sequential()
    model.add(GRU(units=120, activation='relu', input_shape=(n_timesteps,n_features)))
    model.add(Dropout(0.3))
    model.add(Dense(120, activation='relu'))
    model.add(Dense(2, activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=keras.metrics.SpecificityAtSensitivity(sensitivity=0.98))

    model.fit(X_train, Y_train, epochs=epochs, batch_size=batch_size, verbose=verbose)
    
    best_threshold = 0.0
    best_tss = 0.0
    y_pred = model.predict(X_test)
    # evaluate model
    for i in range(1, 1000):

        threshold = i / 1000 # Adjust the threshold as needed
        y_pred_binary = (y_pred > threshold).astype(int)
        confusion = confusion_matrix(Y_test, y_pred_binary)
        tn, fp, fn, tp = confusion.ravel()
        tss = TSS(tp,tn,fp,fn)
        if tss > best_tss:
            best_tss = tss
            best_threshold = i / 1000
        
    
    print(str(X_train.shape)+': GRU Classifier is Done! \n')


    threshold = best_threshold # Adjust the threshold as needed
    y_pred_binary = (y_pred > threshold).astype(int)
    confusion = confusion_matrix(Y_test, y_pred_binary)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "mlp_model.pkl")

    #loaded_mlp_model = joblib.load(data_dir + "mlp_model.pkl")
    
    return output_values

# Results

def save_results(reslut, name):
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/results/"

    with open(data_dir + name + ".pkl", 'wb') as f:
        pickle.dump(reslut, f)

### Rocket

# Rocket Concatenation
rocket_concat_RTS = kfold_training('Rocket', X_train_concat_ZM_RTS, Y_train_concat_ZM_RTS, X_test_concat_ZM, Y_test_concat_ZM, rocket_model, 4)

save_results(rocket_concat_RTS, "Rocket_RusTomekGNI_WithoutC_Concatenation_Results")

rocket_concat_RTA = kfold_training('Rocket', X_train_concat_ZM_RTA, Y_train_concat_ZM_RTA, X_test_concat_ZM, Y_test_concat_ZM, rocket_model, 4)

save_results(rocket_concat_RTA, "Rocket_RusTomekTimeGAN_WithoutC_Concatenation_Results")

rocket_concat_S = kfold_training('Rocket', X_train_concat_ZM_S, Y_train_concat_ZM_S, X_test_concat_ZM, Y_test_concat_ZM, rocket_model, 4)

save_results(rocket_concat_S, "Rocket_GNI_WithoutC_Concatenation_Results")

rocket_concat_A = kfold_training('Rocket', X_train_concat_ZM_A, Y_train_concat_ZM_A, X_test_concat_ZM, Y_test_concat_ZM, rocket_model, 4)

save_results(rocket_concat_A, "Rocket_TimeGAN_WithoutC_Concatenation_Results")

### ShapeDTW

knnts_concat_RTS = kfold_training('ShapeDTW', X_train_concat_ZM_RTS, Y_train_concat_ZM_RTS, X_test_concat_ZM, Y_test_concat_ZM, knnts_model, 4)

save_results(knnts_concat_RTS, "ShapeDTW_RusTomekGNI_WithoutC_Concatenation_Results")

knnts_concat_RTA = kfold_training('ShapeDTW', X_train_concat_ZM_RTA, Y_train_concat_ZM_RTA, X_test_concat_ZM, Y_test_concat_ZM, knnts_model, 4)

save_results(knnts_concat_RTA, "ShapeDTW_RusTomekTimeGAN_WithoutC_Concatenation_Results")

knnts_concat_S = kfold_training('ShapeDTW', X_train_concat_ZM_S, Y_train_concat_ZM_S, X_test_concat_ZM, Y_test_concat_ZM, knnts_model, 4)

save_results(knnts_concat_S, "ShapeDTW_GNI_WithoutC_Concatenation_Results")

knnts_concat_A = kfold_training('ShapeDTW', X_train_concat_ZM_A, Y_train_concat_ZM_A, X_test_concat_ZM, Y_test_concat_ZM, knnts_model, 4)

save_results(knnts_concat_A, "ShapeDTW_TimeGAN_WithoutC_Concatenation_Results")

### TSF

# TSF Concatenation
tsf_concat_RTS = kfold_training('TSF', X_train_concat_ZM_RTS, Y_train_concat_ZM_RTS, X_test_concat_ZM, Y_test_concat_ZM, tsf_model, 4)

save_results(tsf_concat_RTS, "TSF_RusTomekGNI_WithoutC_Concatenation_Results")

tsf_concat_RTA = kfold_training('TSF', X_train_concat_ZM_RTA, Y_train_concat_ZM_RTA, X_test_concat_ZM, Y_test_concat_ZM, tsf_model, 4)

save_results(tsf_concat_RTA, "TSF_RusTomekTimeGAN_WithoutC_Concatenation_Results")

tsf_concat_S = kfold_training('TSF', X_train_concat_ZM_S, Y_train_concat_ZM_S, X_test_concat_ZM, Y_test_concat_ZM, tsf_model, 4)

save_results(tsf_concat_S, "TSF_GNI_WithoutC_Concatenation_Results")

tsf_concat_A = kfold_training('TSF', X_train_concat_ZM_A, Y_train_concat_ZM_A, X_test_concat_ZM, Y_test_concat_ZM, tsf_model, 4)

save_results(tsf_concat_A, "TSF_TimeGAN_WithoutC_Concatenation_Results")

## 2D to 3D

def the_2dto3d(X_train):
    num_attributes = 24
    num_partitions = 5
    num_timestamps = 60
    X_train_concat_ZM_3D = []
    for i in range(0, num_partitions):
        new_3D = np.zeros((X_train[i].shape[0], num_timestamps, num_attributes))

        for j in range(0, X_train[i].shape[0]):
            for m in range(0, num_attributes):
                new_3D[j,:,m] = X_train[i][j,m*num_timestamps:(m+1)*num_timestamps]
        X_train_concat_ZM_3D.append(new_3D)
    return X_train_concat_ZM_3D

X_train_concat_ZM_RTS_3D = the_2dto3d(X_train_concat_ZM_RTS)
X_train_concat_ZM_RTA_3D = the_2dto3d(X_train_concat_ZM_RTA)
del X_train_concat_ZM_RTS
del X_train_concat_ZM_RTA

X_test_concat_ZM_3D = the_2dto3d(X_test_concat_ZM)
del X_test_concat_ZM

X_train_concat_ZM_S_3D = the_2dto3d(X_train_concat_ZM_S)
X_train_concat_ZM_A_3D = the_2dto3d(X_train_concat_ZM_A)
del X_train_concat_ZM_S
del X_train_concat_ZM_A

concat_data_3D = the_2dto3d(concat_data)
del concat_data

### LSTM

# LSTM Concatenation
lstm_concat_RTS = kfold_training('LSTM', X_train_concat_ZM_RTS_3D, Y_train_concat_ZM_RTS, X_test_concat_ZM_3D, Y_test_concat_ZM, lstm_model, 4)

save_results(lstm_concat_RTS, "LSTM_RusTomekGNI_WithoutC_Concatenation_Results")

lstm_concat_RTA = kfold_training('LSTM', X_train_concat_ZM_RTA_3D, Y_train_concat_ZM_RTA, X_test_concat_ZM_3D, Y_test_concat_ZM, lstm_model, 4)

save_results(lstm_concat_RTA, "LSTM_RusTomekTimeGAN_WithoutC_Concatenation_Results")

lstm_concat_S = kfold_training('LSTM', X_train_concat_ZM_S_3D, Y_train_concat_ZM_S, X_test_concat_ZM_3D, Y_test_concat_ZM, lstm_model, 4)

save_results(lstm_concat_S, "LSTM_GNI_WithoutC_Concatenation_Results")

lstm_concat_A = kfold_training('LSTM', X_train_concat_ZM_A_3D, Y_train_concat_ZM_A, X_test_concat_ZM_3D, Y_test_concat_ZM, lstm_model, 4)

save_results(lstm_concat_A, "LSTM_TimeGAN_WithoutC_Concatenation_Results")

raw_concat= kfold_training('LSTM', concat_data_3D, concat_data_labels, concat_data_3D, concat_data_labels, lstm_model, 4)

save_results(raw_concat, "LSTM_Raw_Concatenation_Results")

### 1D-CNN

cnn_concat_RTS = kfold_training('1DCNN', X_train_concat_ZM_RTS_3D, Y_train_concat_ZM_RTS, X_test_concat_ZM_3D, Y_test_concat_ZM, cnn_model, 4)

save_results(cnn_concat_RTS, "1DCNN_RusTomekGNI_WithoutC_Concatenation_Results")

cnn_concat_RTA = kfold_training('1DCNN', X_train_concat_ZM_RTA_3D, Y_train_concat_ZM_RTA, X_test_concat_ZM_3D, Y_test_concat_ZM, cnn_model, 4)

save_results(cnn_concat_RTA, "1DCNN_RusTomekTimeGAN_WithoutC_Concatenation_Results")

cnn_concat_S = kfold_training('1DCNN', X_train_concat_ZM_S_3D, Y_train_concat_ZM_S, X_test_concat_ZM_3D, Y_test_concat_ZM, cnn_model, 4)

save_results(cnn_concat_S, "1DCNN_GNI_WithoutC_Concatenation_Results")

cnn_concat_A = kfold_training('1DCNN', X_train_concat_ZM_A_3D, Y_train_concat_ZM_A, X_test_concat_ZM_3D, Y_test_concat_ZM, cnn_model, 4)

save_results(cnn_concat_A, "1DCNN_TimeGAN_WithoutC_Concatenation_Results")

raw_concat= kfold_training('1DCNN', concat_data_3D, concat_data_labels, concat_data_3D, concat_data_labels, cnn_model, 4)

save_results(raw_concat, "1DCNN_Raw_Concatenation_Results")

### RNN

rnn_concat_RTS = kfold_training('RNN', X_train_concat_ZM_RTS_3D, Y_train_concat_ZM_RTS, X_test_concat_ZM_3D, Y_test_concat_ZM, rnn_model, 4)

save_results(rnn_concat_RTS, "RNN_RusTomekGNI_WithoutC_Concatenation_Results")

rnn_concat_RTA = kfold_training('RNN', X_train_concat_ZM_RTA_3D, Y_train_concat_ZM_RTA, X_test_concat_ZM_3D, Y_test_concat_ZM, rnn_model, 4)

save_results(rnn_concat_RTA, "RNN_RusTomekTimeGAN_WithoutC_Concatenation_Results")

rnn_concat_S = kfold_training('RNN', X_train_concat_ZM_S_3D, Y_train_concat_ZM_S, X_test_concat_ZM_3D, Y_test_concat_ZM, rnn_model, 4)

save_results(rnn_concat_S, "RNN_GNI_WithoutC_Concatenation_Results")

rnn_concat_A = kfold_training('RNN', X_train_concat_ZM_A_3D, Y_train_concat_ZM_A, X_test_concat_ZM_3D, Y_test_concat_ZM, rnn_model, 4)

save_results(rnn_concat_A, "RNN_TimeGAN_WithoutC_Concatenation_Results")

raw_concat= kfold_training('RNN', concat_data_3D, concat_data_labels, concat_data_3D, concat_data_labels, rnn_model, 4)

save_results(raw_concat, "RNN_Raw_Concatenation_Results")

### GRU

gru_concat_RTS = kfold_training('GRU', X_train_concat_ZM_RTS_3D, Y_train_concat_ZM_RTS, X_test_concat_ZM_3D, Y_test_concat_ZM, gru_model, 4)

save_results(gru_concat_RTS, "GRU_RusTomekGNI_WithoutC_Concatenation_Results")

gru_concat_RTA = kfold_training('GRU', X_train_concat_ZM_RTA_3D, Y_train_concat_ZM_RTA, X_test_concat_ZM_3D, Y_test_concat_ZM, gru_model, 4)

save_results(gru_concat_RTA, "GRU_RusTomekTimeGAN_WithoutC_Concatenation_Results")

gru_concat_S = kfold_training('GRU', X_train_concat_ZM_S_3D, Y_train_concat_ZM_S, X_test_concat_ZM_3D, Y_test_concat_ZM, gru_model, 4)

save_results(gru_concat_S, "GRU_GNI_WithoutC_Concatenation_Results")

gru_concat_A = kfold_training('GRU', X_train_concat_ZM_A_3D, Y_train_concat_ZM_A, X_test_concat_ZM_3D, Y_test_concat_ZM, gru_model, 4)

save_results(gru_concat_A, "GRU_TimeGAN_WithoutC_Concatenation_Results")

raw_concat= kfold_training('GRU', concat_data_3D, concat_data_labels, concat_data_3D, concat_data_labels, gru_model, 4)

save_results(raw_concat, "GRU_Raw_Concatenation_Results")

# Comparison 

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/results/"
model_name = ['LSTM','1DCNN','RNN','GRU']
ousampling_name = ['RusTomekGNI','RusTomekTimeGAN','GNI','TimeGAN']

values = np.array([])
names = []
count = 0
for name in model_name:
    for s_name in ousampling_name:
        with open(data_dir + name + '_' + s_name + '_WithoutC_Concatenation_Results' + ".pkl", 'rb') as f:
            if count == 0:
                values=np.array(pickle.load(f))
            else:
                values = np.append(values, np.array(pickle.load(f)), axis=0)
            names.append( name+ ' (' + s_name + ')' )
            count += 1

def compare_results(names, values, num):
    train_test_counts = 1
    np.printoptions(precision=4, suppress=True)
    for i in range(0, int(values.shape[0]/num)):
        print("P_Train = "+ str(values[i,0]) + " & " + "P_Test = " + str(values[i,1]))
        for j in range(0, num):
            print(names[j] + ' :' +  ' TP={:.0f}'.format(values[j*train_test_counts+i,2]) + ' FN={:.0f}'.format(values[j*train_test_counts+i,3]) + ' FP={:.0f}'.format(values[j*train_test_counts+i,4])
                 + ' TN={:.0f}'.format(values[j*train_test_counts+i,5]) + ' TSS={:.3f}'.format(values[j*train_test_counts+i,6]) + ' HSS1={:.3f}'.format(values[j*train_test_counts+i,7]) + ' HSS2={:.3f}'.format(values[j*train_test_counts+i,8])
                 + ' GSS={:.3f}'.format(values[j*train_test_counts+i,9]) + ' Recall={:.3f}'.format(values[j*train_test_counts+i,10]) + ' Precision={:.3f}'.format(values[j*train_test_counts+i,11]))
        print('\n')


compare_results(names, values, 16)
